Move `showAddSubscribersModal`, `setShowAddSubscribersModal` and `addSubscribersCallback` into `SubscribersPageProvider`
