+++ client/my-sites/subscribers/components/add-subscribers-modal/add-subscribers-modal.tsx
@@ -5,6 +5,7 @@ import { AddSubscriberForm } from '@automattic/subscriber';
 import { Modal } from '@wordpress/components';
 import { useI18n } from '@wordpress/react-i18n/';
 import { useTranslate } from 'i18n-calypso';
+import { useSubscribersPage } from 'calypso/my-sites/subscribers/components/subscribers-page/subscribers-page-context';
 import { useState } from 'react';
 import { LoadingBar } from 'calypso/components/loading-bar';
 import './style.scss';
@@ -12,21 +13,16 @@ import './style.scss';
 type AddSubscribersModalProps = {
 	siteId: number;
 	siteTitle: string;
-	showModal: boolean;
-	onClose: () => void;
-	onAddFinished: () => void;
 };
 
 const AddSubscribersModal = ( {
 	siteId,
 	siteTitle,
-	showModal,
-	onClose,
-	onAddFinished,
 }: AddSubscribersModalProps ) => {
 	const translate = useTranslate();
 	const locale = useLocale();
 	const { hasTranslation } = useI18n();
+	const { showAddSubscribersModal, setShowAddSubscribersModal, addSubscribersCallback } = useSubscribersPage();
 
 	const modalTitle =
 		locale.startsWith( 'en' ) || hasTranslation( 'Add subscribers to %s' )
@@ -40,17 +36,17 @@ const AddSubscribersModal = ( {
 	const onImportStarted = ( hasFile: boolean ) => setIsUploading( hasFile );
 	const onImportFinished = () => {
 		setIsUploading( false );
-		onAddFinished();
+		addSubscribersCallback();
 	};
 
-	if ( ! showModal ) {
+	if ( ! showAddSubscribersModal ) {
 		return null;
 	}
 
 	return (
 		<Modal
 			title={ modalTitle as string }
-			onRequestClose={ onClose }
+			onRequestClose={ () => setShowAddSubscribersModal( false ) }
 			overlayClassName="add-subscribers-modal"
 		>
 			{ isUploading && (
