+++ client/my-sites/subscribers/components/subscriber-list-container/subscriber-list-container.tsx
@@ -13,13 +13,11 @@ import './style.scss';
 type SubscriberListContainerProps = {
 	onClickView: ( subscriber: Subscriber ) => void;
 	onClickUnsubscribe: ( subscriber: Subscriber ) => void;
-	setShowAddSubscribersModal: React.Dispatch< React.SetStateAction< boolean > >;
 };
 
 const SubscriberListContainer = ( {
 	onClickView,
 	onClickUnsubscribe,
-	setShowAddSubscribersModal,
 }: SubscriberListContainerProps ) => {
 	const { grandTotal, total, perPage, page, pageChangeCallback, searchTerm } = useSubscribersPage();
 	useRecordSearch();
@@ -37,10 +35,7 @@ const SubscriberListContainer = ( {
 					{ total ? (
 						<SubscriberList onView={ onClickView } onUnsubscribe={ onClickUnsubscribe } />
 					) : (
-						<NoSearchResults
-							searchTerm={ searchTerm }
-							setShowAddSubscribersModal={ setShowAddSubscribersModal }
-						/>
+						<NoSearchResults searchTerm={ searchTerm } />
 					) }
 
 					<Pagination
