+++ client/my-sites/subscribers/components/subscribers-header/subscribers-header.tsx
@@ -8,15 +8,10 @@ import { SubscribersHeaderPopover } from '../subscribers-header-popover';
 type SubscribersHeaderProps = {
 	selectedSiteId: number | undefined;
 	navigationItems: Item[];
-	setShowAddSubscribersModal: React.Dispatch< React.SetStateAction< boolean > >;
 };
 
-const SubscribersHeader = ( {
-	navigationItems,
-	selectedSiteId,
-	setShowAddSubscribersModal,
-}: SubscribersHeaderProps ) => {
-	const { grandTotal } = useSubscribersPage();
+const SubscribersHeader = ( { navigationItems, selectedSiteId }: SubscribersHeaderProps ) => {
+	const { grandTotal, setShowAddSubscribersModal } = useSubscribersPage();
 
 	return (
 		<FixedNavigationHeader navigationItems={ navigationItems }>
